requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
openai==1.109.1
azure-search-documents==11.6.0b12
azure-cosmos==4.9.1
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-monitor-opentelemetry==1.8.1
opentelemetry-api==1.36.0
pydantic==2.11.9
pydantic-settings==2.11.0
httpx==0.28.1
